+++ packages/patrol_devtools_extension/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Regenerate API contracts for the new Android `ConfigureRequest` accessibility flag. (#3227)
 - Regenerate API contracts (adds `AndroidTakeNativeScreenshotRequest`). (#3222)
+- Regenerate API contracts (adds `IOSTakeNativeScreenshotRequest`). (#3271)
 
 ## 0.4.1